expect-elixir v0.0.3 Expect

Tiny TCL/Expect-ish interface for the excellent Porcelain library.

Types

pattern :: binary | Regex.t

Functions

close(process)

Specs

close(%{}) :: :ok

Close a spawned process.

do_expect(map, timeout, expects)

Specs

do_expect(%{}, pos_integer, [{pattern, function}]) ::
  {:ok, any} |
  {:error, :exit, non_neg_integer, list} |
  {:error, :etimedout}

Called by the expect/2, expect/3, and expect/4 macros.

send(process, data)

Specs

send(%{}, binary) :: :ok

Send data to a spawned process.

spawn(command)

Specs

spawn(String.t) :: {:ok, %{}}

Spawn a process for command.

Macros

exp_close(process)

An alias for Expect.close/1.

exp_send(process, data)

An alias for Expect.send/2.

exp_spawn(command)

An alias for Expect.spawn/1.

expect(process, list)

Match incoming data and execute the corresponding block.

Matched data is accessed from within a block via the buffer variable.

If no more specific match is found, the first block matching _ is executed. If no data is received within the timeout period, no block matching _ will be executed.

If no match is found within the timeout period, the first block matching :timeout is executed, or else the first block matching :default is executed. Otherwise, {:error, :etimedout} is returned.

Examples

iex> expect spawned_process do
  ~r/>$/ ->
    :match

  "a string data may contain" ->
    buffer

after
  2_000 ->
    :not_what_i_expected
end
{:ok, :match}

expect(process, timeout, pattern, fun \\ quote() do &(&1) end)

Match incoming data on pattern and execute fun/1.

Matched data is accessed from within fun/1 with buffer.

If pattern is _ and data is received, fun is executed. If pattern is _ and no data is received within the timeout period, {:error, :etimedout} is returned.

If no match is found within the timeout period and pattern is :timeout, fun is executed. If no match is found within the timeout period and pattern is :default, fun is executed. Otherwise, {:error, :etimedout} is returned.

Examples

iex> expect(spawned_process, 2_000, ~r/>$/)
{:ok, ["some data

“, “some more data>”]}

iex> expect(spawned_process, 2_000, _, fn _ -> :got_something end)
{:ok, :got_something}