expect-elixir v0.0.4 Expect

Tiny TCL/Expect-ish interface for the excellent Porcelain library.

Types

expect_error ::
  {:error, :etimedout} |
  {:error, :exit, non_neg_integer, binary}

pattern :: binary | Regex.t

process :: %{pid: pid}

Functions

close(process)

Specs

close(process) :: :ok

Close a spawned process.

exp_close(process)

Calls Expect.close/1. Imported with use Expect.

exp_send(process, data)

Calls Expect.send/2. Imported with use Expect.

exp_spawn(command)

Calls Expect.spawn/1. Imported with use Expect.

expect(map, timeout \\ 10000, pattern_or_fun)

Specs

expect(process, pos_integer, pattern) ::
  nil |
  expect_error

expect(process, pos_integer, function) ::
  {:ok, any} |
  expect_error

Match and process incoming data from spawned process.

expect behaves differently contingent on whether it is given a function or a bare pattern.

When a bare binary or regex pattern is provided, expect returns one of the following.

on match	nil

on timeout	{:timeout, buffer} or {:default, buffer}

on exit	{:default, buffer}

This behavior is useful for when you don’t care about the incoming data so much as the fact that it arrived (and matched your pattern).

When a function is provided, expect keeps a buffer of unmatched data which is passed to fun when: new data arrives; the timeout period expires; the spawned process exits. In each case, the value(s) passed to fun take(s) one of the following forms.

on match	{:data, buffer}

on timeout	{:timeout, buffer}, then {:default, buffer}

on exit	{:default, buffer}

Examples

iex> expect spawned_process, "# "
nil

iex> expect spawned_process, fn {event, buffer} ->
  cond do
    buffer =~ ~r/>$/ ->
      :match

    buffer =~ "a string" ->
      buffer

    event == :timeout ->
      :it_timed_out
  end
end
{:ok, :match}

iex> expect spawned_process, fn {_, buffer} -> buffer end
{:ok, ["some data\r\n", "some more data\r\n"]}

iex> expect process_without_data, fn {:data, buffer} -> buffer end
{:error, :etimedout}

iex> expect spawned_process, fn buffer ->
  cond do
    buffer =~ "non-matching pattern" ->
      buffer
  end
end
{:error, :etimedout}

iex> expect spawned_process, fn {:timeout, _} -> :it_timed_out end
{:ok, :it_timed_out}

send(process, data)

Specs

send(process, binary) :: :ok

Send data to a spawned process.

spawn(command)

Specs

spawn(String.t) :: {:ok, process}

Spawn a process for command.